Update main.py
Init the pygame mixer
Add debug tools with is settings
Add load sound method (return a class if the sound is not available).
Load all assets, create players and the ball with load_assets methods.
Check if a player get the maxixum score.
Create reset methods when a player get the maximum score.
Add text rendering (start the level or reset the game)